+++ cmd-attach-session.c
@@ -76,7 +76,7 @@ cmd_attach_session(struct cmdq_item *item, const char *tflag, int dflag,
 	}
 
 	if (!server_acl_attach_session(c)) {
-		cmdq_error(item, TMUX_ACL_LOG " ACL attach session error");
+		cmdq_error(item, TMUX_ACL_LOG, " ACL attach session error");
 		return (CMD_RETURN_ERROR);
 	}
