+++ tmux.h
@@ -3190,30 +3190,19 @@ struct window_pane *spawn_pane(struct spawn_context *, char **);
 /* regsub.c */
 char		*regsub(const char *, const char *, const char *, int);
 
-#define TMUX_ACL_LOG "[access control list]"
 
 /* server-acl.c */
-
+#define TMUX_ACL_LOG "[access control list]"
 void server_acl_init(void);
-
 void server_acl_user_allow(uid_t uid, int owner);
-
 void server_acl_user_deny(uid_t uid);
-
 struct acl_user* server_acl_user_find(uid_t uid);
-
 int server_acl_check_host(uid_t uid);
-
 int server_acl_accept_validate(int newf, struct clients clients);
-
 int server_acl_attach_session(struct client *c);
-
 struct passwd;
-
 void server_acl_user_allow_write(struct passwd* user_data);
-
 void server_acl_user_deny_write(struct passwd* user_data);
-
 void server_acl_client_fail(const char* message, ...);
 
 #endif /* TMUX_H */
